add offline mode
