+++ package.json
@@ -7,8 +7,7 @@
     "build": "webpack",
     "start": "webpack-dev-server --open",
     "eslint": "eslint src/**/*.js",
-    "test": "npm run eslint",
-    "serv": "start json-server --watch server/data.json --port 3003"
+    "test": "npm run eslint"
   },
   "devDependencies": {
     "eslint": "6.1.0",
