+++ src/api.js
@@ -86,4 +86,4 @@ export default class API {
         throw err;
       });
   }
-};
+}
