+++ src/components/statistics.js
@@ -40,7 +40,6 @@ export default class Stats extends AbstractComponent {
   }
 
   updateData(arr) {
-    console.log(arr);
     this._prepareData(arr);
     this.moneyChart.data.datasets[0].data = this.prices;
     this.moneyChart.data.labels = this.labels;
@@ -52,7 +51,6 @@ export default class Stats extends AbstractComponent {
     this.moneyChart.update();
     this.transportChart.update();
     this.timeChart.update();
-    console.log(this.moneyChart.data);
   }
 
   _prepareData(arr) {
@@ -79,7 +77,7 @@ export default class Stats extends AbstractComponent {
 
       const duration = Math.round(moment.duration(moment(point.dateTo).diff(moment(point.dateFrom))).asHours());
       const curIndexPlace = this.placesNames.findIndex((elem) => elem === point.destination.name);
-      if (curIndexPlace === -1) {
+      if (curIndexPlace === -1 && duration > 0) {
         this.placesNames.push(point.destination.name);
         this.placesTime.push(duration);
       } else {
